requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
streamlit==1.41.3
langchain==1.1.3
langchain-openai==0.2.16
langchain-core==0.3.32
langgraph==0.2.69
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
azure-identity==1.19.0
httpx==0.28.1
tavily-python==0.5.0
pymilvus==2.5.4
python-multipart==0.0.20
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
rank-bm25==0.2.2
jieba==0.42.1
